LayoutTests/http/tests/xmlhttprequest/access-control-basic-get-fail-non-simple-expected.txt

CONSOLE MESSAGE: line 28: XMLHttpRequest cannot load http://localhost:8000/xmlhttprequest/resources/access-control-basic-get-fail-non-simple.cgi. Preflight response is not successful
PASS: Exception thrown. Cross-domain access was denied in 'send'. [ A network error occurred.].
